fix auth and routes

+++ src/components/adminPackages/adminPackages.js
@@ -2,6 +2,8 @@ import React, { useState, useEffect, useContext } from "react";
 import "./adminPackages.css";
 import axios from "axios";
 import Cookies from "universal-cookie";
+import Swal from 'sweetalert2'
+
 const cookies = new Cookies();
 const user = cookies.get("data")
 function AdminPackages() {
@@ -122,6 +124,13 @@ function AdminPackages() {
         console.log(err);
       });
   };
+  const swalWithBootstrapButtons = Swal.mixin({
+    customClass: {
+        confirmButton: 'btn btn-success',
+        cancelButton: 'btn btn-danger'
+    },
+    buttonsStyling: false
+})
 
   useEffect(() => {
     getAllPackages();
@@ -131,7 +140,27 @@ function AdminPackages() {
   }, [allPackages]);
 
   function deleteThePackage(id) {
-    deletePackage(id);
+
+    swalWithBootstrapButtons
+            .fire({
+            title: 'Are you sure?',
+            text: "You won't be able to revert this!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonText: 'Yes, delete it!',
+            cancelButtonText: 'No, cancel!',
+            reverseButtons: true
+        })
+            .then((result) => {
+                if (result.isConfirmed) { //true =yes
+                  deletePackage(id);
+                  swalWithBootstrapButtons.fire('Deleted!', 'Your file has been deleted.', 'success')
+                } else if ( //false cancel
+                /* Read more about handling dismissals below */
+                result.dismiss === Swal.DismissReason.cancel) {
+                    swalWithBootstrapButtons.fire('Cancelled', 'Your imaginary file is safe :)', 'error')
+                }
+            })
   }
 
   function acceptThePackage(id) {
